Fix Gson date format in Android Marshmallow
[Description]
When opening the app on Android Marshmallow, a FC happens.

[Root cause]
The date format "XXX" doesn't seem to be valid in Android Marshmallow. In
most recent versions, like Nougat, that format works fine.

[Solution]
Use the format "Z" instead of "XXX" to make it compatible with Android
Marshmallow.

[Test]
0) Make sure the device is running Android Marshmallow.
1) Launch the app.

After step 1, the app should be launched correctly without crashing.

Signed-off-by: Crístian Deives <cdeives@motorola.com>