Keep "Add bus" dialog button disabled when rotating screen
[Description]
The button "OK" in the "Add bus" dialog should be enabled when there's
any valid text in the text field, and disabled otherwise. If the user
types something in the text field and rotates the screen, the button
state changes from enabled to disabled.

[Root cause]
The button state is initialized to "disabled" every time a new dialog is
created. This doesn't take into account an initialization due to
rotation change.

[Solution]
During the dialog initialization, set the button state according to the
text field content.

[Test]
1) Tap the button "+" on the first tab;
2) Type anything in the new dialog;
3) Rotate the screen;

After step (3), the button "OK" should still be enabled.

Signed-off-by: Crístian Deives <cdeives@motorola.com>
Trello-card: https://trello.com/c/jb39giDT