Fix FC when checking switch when GMS not installed
[Description]
When the user doesn't have Google Play Services installed and tries to
check the switch "Detect movement automatically", a FC happens.

[Root cause]
When the user checks the switch, the app detects that the GMS connection
failed and tries to "rollback" the activity detection service, which has
just been activated. However, the app didn't assume that, when disabling
the service, the GMS connection might not be valid.

[Solution]
Check if the GMS connection is established before trying to deactivate
it.

[Test]
0) Make sure the device doesn't have Google Play Services installed;
1) Tap the switch "Detect movement automatically";

After step (1), the app shouldn't crash and the switch should still be
unchecked.

Signed-off-by: Crístian Deives <cdeives@motorola.com>