Keep "Detect movement automatically" switch checked when changing tabs
[Description]
When the user checks the switch "Detect movement automatically" and go
to the tab "View" and then back to the tab "Track", the switch becomes
unchecked and toast error message is displayed.

[Root cause]
The presenter wasn't correctly reset on TrackBusFragment::onCreateView,
so when the user returned to the tab "View", the leaked presenter still
didn't know about the bus loaders on the spinner (because the fragment
hasn't fetched the current data yet). That causes the fragment to think
that the user is trying to check the switch without selecting a bus
number, then it unchecks the switch and displays an error message.

[Solution]
Set the presenter to null on TrackBusFragment::onCreateView. The code
already ignores validation due to the switch change when the presenter
is null.

[Test]
0. Make sure there's at least one bus number on the spinner;
1. Check the switch "Detect movement automatically";
2. Go to the tab "View";
3. Go back to the tab "Track";

After step (3), the switch from step (1) should still be checked and no
error message should be displayed.

Signed-off-by: Crístian Deives <cdeives@motorola.com>
Trello-card: https://trello.com/c/smmX5ReR