Fix FC when launching dialog to create bus number
[Description]
When the user taps the "+" button to create a new bus, a FC happens.

[Root cause]
The dialog was trying to access a UI component with findViewByID but the
component is actually inside the dialog object returned by getDialog().
As the component can't be found, a NPE is thrown.

[Solution]
Access the UI component with getDialog().findViewByID() instead.

[Test]
1. Tap the "+" button.

After step (1), a dialog should be displayed and no FC should be seen.

Signed-off-by: Crístian Deives <cdeives@motorola.com>
Trello-card: https://trello.com/c/HHkvIAg7