Do not allow bus number to be created while tracking bus
[Description]
If the user is tracking the bus (or if the activity detection service is
active) and a new bus number is created, the spinner is updated with the
new value and it becomes enabled.

[Root cause]
The "+" button is enabled all the time, even in situations where it
doesn't make sense to create a new bus number.

[Solution]
Disable the "+" button when the user is tracking the bus (of if the
activity detection service is active).

[Test]
1. Check the switch "Detect movement automatically" (or tap the button
"Enter bus");
2. Tap the "+" button.

After step (2), nothing should happen; the dialog to create the new bus
number should not be displayed.

Signed-off-by: Crístian Deives <cdeives@motorola.com>
Trello-card: https://trello.com/c/EGaEoonl